+++ graylog2-web-interface/src/components/authentication/directoryServices/activeDirectory/BackendEdit.jsx
@@ -8,8 +8,9 @@ import { getEnterpriseGroupSyncPlugin } from 'util/AuthenticationService';
 import WizardPageHeader from './WizardPageHeader';
 import { HELP, AUTH_BACKEND_META } from './BackendCreate';
 
-import { prepareInitialValues, handleSubmit } from '../ldap/BackendEdit';
+import { prepareInitialValues } from '../ldap/BackendEdit';
 import BackendWizard from '../BackendWizard';
+import handleUpdate from '../HandleUpdate';
 
 type Props = {
   authenticationBackend: DirectoryServiceBackend,
@@ -56,7 +57,19 @@ const BackendEdit = ({ authenticationBackend, initialStepKey }: Props) => {
     backendHasPassword: authenticationBackend.config.systemUserPassword.isSet,
     backendGroupSyncIsActive: !!initialValues.synchronizeGroups,
   };
-  const _handleSubmit = (payload, formValues, serviceType, shouldUpdateGroupSync) => handleSubmit(payload, formValues, authenticationBackend.id, !!initialValues.synchronizeGroups, serviceType, shouldUpdateGroupSync);
+  const _handleSubmit = (
+    payload,
+    formValues,
+    serviceType,
+    shouldUpdateGroupSync,
+  ) => handleUpdate(
+    payload,
+    formValues,
+    authenticationBackend.id,
+    !!initialValues.synchronizeGroups,
+    serviceType,
+    shouldUpdateGroupSync,
+  );
 
   return (
     <DocumentTitle title="Edit Active Directory Authentication Service">
