+++ graylog2-web-interface/src/components/authentication/directoryServices/ldap/BackendEdit.jsx
@@ -2,17 +2,16 @@
 import * as React from 'react';
 import * as Immutable from 'immutable';
 
-import { AuthenticationActions } from 'stores/authentication/AuthenticationStore';
-import type { DirectoryServiceBackend, WizardSubmitPayload } from 'logic/authentication/directoryServices/types';
+import type { DirectoryServiceBackend } from 'logic/authentication/directoryServices/types';
 import { getEnterpriseGroupSyncPlugin } from 'util/AuthenticationService';
 import { DocumentTitle, Spinner } from 'components/common';
-import UserNotification from 'util/UserNotification';
 
 import { AUTH_BACKEND_META, HELP } from './BackendCreate';
 import WizardPageHeader from './WizardPageHeader';
 
 import type { WizardFormValues } from '../BackendWizard/BackendWizardContext';
 import BackendWizard from '../BackendWizard';
+import handleUpdate from '../HandleUpdate';
 
 type Props = {
   authenticationBackend: DirectoryServiceBackend,
@@ -58,28 +57,6 @@ const _optionalWizardProps = (initialStepKey: ?string) => {
   return props;
 };
 
-export const handleSubmit = (payload: WizardSubmitPayload, formValues: WizardFormValues, backendId: string, backendGroupSyncIsActive: boolean, serviceType: string, shouldUpdateGroupSync: ?boolean = true) => {
-  const enterpriseGroupSyncPlugin = getEnterpriseGroupSyncPlugin();
-  const notifyOnSuccess = () => UserNotification.success('Authentication service was updated successfully.', 'Success');
-  const notifyOnError = (error) => UserNotification.error(`Updating authentication service failed with status: ${error}`, 'Error');
-
-  return AuthenticationActions.update(backendId, {
-    ...payload,
-    id: backendId,
-  }).then((result) => {
-    if (enterpriseGroupSyncPlugin && shouldUpdateGroupSync) {
-      return enterpriseGroupSyncPlugin.actions.onDirectoryServiceBackendUpdate(backendGroupSyncIsActive, formValues, backendId, serviceType).then(notifyOnSuccess);
-    }
-
-    notifyOnSuccess();
-
-    return result;
-  }).catch((error) => {
-    notifyOnError(error);
-    throw error;
-  });
-};
-
 const BackendEdit = ({ authenticationBackend, initialStepKey }: Props) => {
   const enterpriseGroupSyncPlugin = getEnterpriseGroupSyncPlugin();
   const {
@@ -108,7 +85,19 @@ const BackendEdit = ({ authenticationBackend, initialStepKey }: Props) => {
     backendHasPassword: authenticationBackend.config.systemUserPassword.isSet,
     backendGroupSyncIsActive: !!initialValues.synchronizeGroups,
   };
-  const _handleSubmit = (payload, formValues, serviceType, shouldUpdateGroupSync) => handleSubmit(payload, formValues, authenticationBackend.id, !!initialValues.synchronizeGroups, serviceType, shouldUpdateGroupSync);
+  const _handleSubmit = (
+    payload,
+    formValues,
+    serviceType,
+    shouldUpdateGroupSync,
+  ) => handleUpdate(
+    payload,
+    formValues,
+    authenticationBackend.id,
+    !!initialValues.synchronizeGroups,
+    serviceType,
+    shouldUpdateGroupSync,
+  );
 
   return (
     <DocumentTitle title="Edit LDAP Authentication Service">
